fixed not existing files
files shouldn't be created anymore if they dont exist

+++ client.py
@@ -72,6 +72,12 @@ def main():
             mode = input("Enter transfer mode to be used ('netascii' or 'octet'): ")
             file_name = os.path.basename(file_path)
             send_request(sock, server_address, file_name, mode, is_write=False)
+            
+            if not os.path.exists(file_path):
+                print('ERROR: ' + ERROR_CODE[1])
+                print(f"Could not find file: '{filename}'")
+                continue
+
             try:
                 file = open(file_path, "wb")
             except FileNotFoundError:
@@ -90,6 +96,12 @@ def main():
             mode = input("Enter transfer mode to be used ('netascii' or 'octet'): ")
             server_filename = os.path.basename(server_filename)
             send_request(sock, server_address, server_filename, mode, is_write=True)
+
+            if not os.path.exists(file_path):
+                print('ERROR: ' + ERROR_CODE[1])
+                print(f"Could not find file: '{filename}'")
+                continue
+
 
             try:
                 file = open(filename, "rb")
@@ -128,6 +140,7 @@ def main():
 
                     send_data(sock, server, seq_number + 1, file_block)
                     if len(file_block) < BLOCK_SIZE:
+                        print(f"\n{operation.capitalize()} completed successfully.")
                         break
                 elif opcode == OPCODE['ERROR']:
                     error_code = int.from_bytes(data[2:4], byteorder='big')
@@ -140,7 +153,6 @@ def main():
         finally:
             file.close()
 
-        print(f"\n{operation.capitalize()} completed successfully.")
 
     sock.close()
